python_automation_framework/requirements.txt
behave==1.2.6
selenium==4.18.0
allure-behave==2.13.5
requests==2.32.3
pymysql==1.1.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
webdriver-manager==4.0.2
configparser==6.0.0
jsonschema>=4.0.0